+++ lib/input/constructor.go
@@ -226,6 +226,7 @@ const (
 	TypeMQTT              = "mqtt"
 	TypeNanomsg           = "nanomsg"
 	TypeNATS              = "nats"
+	TypeNATSJetStream     = "nats_jetstream"
 	TypeNATSStream        = "nats_stream"
 	TypeNSQ               = "nsq"
 	TypePulsar            = "pulsar"
@@ -283,6 +284,7 @@ type Config struct {
 	MQTT              reader.MQTTConfig            `json:"mqtt" yaml:"mqtt"`
 	Nanomsg           reader.ScaleProtoConfig      `json:"nanomsg" yaml:"nanomsg"`
 	NATS              reader.NATSConfig            `json:"nats" yaml:"nats"`
+	NATSJetStream     NATSJetStreamConfig          `json:"nats_jetstream" yaml:"nats_jetstream"`
 	NATSStream        reader.NATSStreamConfig      `json:"nats_stream" yaml:"nats_stream"`
 	NSQ               reader.NSQConfig             `json:"nsq" yaml:"nsq"`
 	Plugin            interface{}                  `json:"plugin,omitempty" yaml:"plugin,omitempty"`
@@ -341,6 +343,7 @@ func NewConfig() Config {
 		MQTT:              reader.NewMQTTConfig(),
 		Nanomsg:           reader.NewScaleProtoConfig(),
 		NATS:              reader.NewNATSConfig(),
+		NATSJetStream:     NewNATSJetStreamConfig(),
 		NATSStream:        reader.NewNATSStreamConfig(),
 		NSQ:               reader.NewNSQConfig(),
 		Plugin:            nil,