+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - The `http_server` input now adds path parameters (`/{foo}/{bar}`) to the metadata of ingested messages.
 - The `stdout` output now has a `codec` field.
 - New Bloblang methods `format_timestamp_strftime` and `parse_timestamp_strptime`.
+- New experimental `nats_jetstream` input and output.
 
 ### Fixed